+++ src/Application/Query/Agenda/HappeningViewQuery.php
@@ -12,9 +12,13 @@ namespace Proximum\Vimeet\Application\Query\Agenda;
 
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Happening;
+use Proximum\Vimeet\Domain\Model\User;
 
 class HappeningViewQuery
 {
+    /** @var User */
+    public $user;
+
     /** @var Happening */
     public $happening;
 
@@ -24,8 +28,9 @@ class HappeningViewQuery
     /** @var string */
     public $locale;
 
-    public function __construct(Happening $happening, Event $event, $locale)
+    public function __construct(User $user, Happening $happening, Event $event, $locale)
     {
+        $this->user = $user;
         $this->happening = $happening;
         $this->event = $event;
         $this->locale = $locale;