+++ src/Ui/Bundle/EventBundle/Controller/Happening/HappeningWebinarAction.php
@@ -1,0 +1,102 @@
+<?php
+
+/*
+ * This file is part of the Proximum Vimeet project.
+ *
+ * Copyright (C) Proximum
+ *
+ * @author Elao <contact@elao.com>
+ */
+
+namespace Proximum\Vimeet\Ui\Bundle\EventBundle\Controller\Happening;
+
+use Proximum\Vimeet\Application\Adapter\AuthorizationCheckerAdapterInterface;
+use Proximum\Vimeet\Application\Adapter\CommandBusInterface;
+use Proximum\Vimeet\Application\Adapter\QueryBusInterface;
+use Proximum\Vimeet\Application\Command\Happening\Webinar\StartWebinarSessionCommand;
+use Proximum\Vimeet\Application\Query\Happening\Webinar\CanAccessToWebinar;
+use Proximum\Vimeet\Application\Query\Happening\Webinar\GetWebinarViewQuery;
+use Proximum\Vimeet\Application\View\Happening\WebinarView;
+use Proximum\Vimeet\Domain\Model\Happening;
+use Proximum\Vimeet\Domain\Model\Sheet;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\ParamConverter\EventDomain;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\Security\Happening\ParticipationVoter;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\Security\SheetVoter;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\ValueResolver\UserDomain;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Symfony\Component\Templating\EngineInterface;
+
+class HappeningWebinarAction
+{
+    /** @var AuthorizationCheckerAdapterInterface */
+    private $authorizationCheckerAdapter;
+
+    /** @var CanAccessToWebinar */
+    private $canAccessToWebinar;
+
+    /** @var CommandBusInterface */
+    private $commandBus;
+
+    /** @var EngineInterface */
+    private $engine;
+
+    /** @var QueryBusInterface */
+    private $queryBus;
+
+    public function __construct(
+        AuthorizationCheckerAdapterInterface $authorizationCheckerAdapter,
+        CanAccessToWebinar $canAccessToWebinar,
+        CommandBusInterface $commandBus,
+        EngineInterface $engine,
+        QueryBusInterface $queryBus
+    ) {
+        $this->authorizationCheckerAdapter = $authorizationCheckerAdapter;
+        $this->canAccessToWebinar = $canAccessToWebinar;
+        $this->commandBus = $commandBus;
+        $this->engine = $engine;
+        $this->queryBus = $queryBus;
+    }
+
+    public function __invoke(
+        Request $request,
+        EventDomain $eventDomain,
+        Sheet $sheet,
+        Happening $happening,
+        UserDomain $userDomain
+    ): Response {
+        $event = $eventDomain->getEvent();
+        $user = $userDomain->getUser();
+
+        if (!$this->authorizationCheckerAdapter->isGranted('IS_AUTHENTICATED_REMEMBERED')
+            || !$this->authorizationCheckerAdapter->isGranted('PERMISSION_HAPPENING_ACCESS', $event)
+            || !$this->authorizationCheckerAdapter->isGranted(SheetVoter::EDIT, $sheet)
+            || !$this->authorizationCheckerAdapter->isGranted(ParticipationVoter::PARTICIPATE, $sheet)
+            || !$this->canAccessToWebinar->isSatisfiableBy($happening, $user)
+            || $happening->getEvent() !== $event
+            || $sheet->getEvent() !== $event
+        ) {
+            throw new AccessDeniedException('Access denied to this happening');
+        }
+
+        $this->commandBus->handle(new StartWebinarSessionCommand($happening));
+
+        /** @var WebinarView $webinarView */
+        $webinarView = $this->queryBus->handle(new GetWebinarViewQuery($happening, $user, $request->getLocale()));
+
+        return new Response(
+            $this->engine->render(
+                $webinarView->isSpeaker
+                    ? '@Event/Happening/webinar-speaker.html.twig'
+                    : '@Event/Happening/webinar-viewer.html.twig',
+                [
+                    'event' => $event,
+                    'sheet' => $sheet,
+                    'userCompleteName' => $user->getAccount()->getCompleteName(),
+                    'webinarView' => $webinarView,
+                ]
+            )
+        );
+    }
+}